requirements.txt
langchain==0.3.3
flask==3.0.3
langchain_groq==0.2.2
langchain-community==0.3.2
langchain_huggingface==0.1.0
mysql==0.0.3
faiss-cpu==1.9.0.post1
mysql-connector-python==9.1.0
python-dotenv==1.0.1